ccl::hash-pname loses when given a non-simple-string
find-all-symbols did no checking and didn't massage.  added a
ccl::ensure-simple-string to both check and massage.

make-symbol checked that the arg was a 'simple-string using
#'require-type.  replacing that with ccl::ensure-simple-string both
checks and massages.

ccl::pkg-arg (called by a bunch of symbol iterator macros) did check
that if the arg was a string, that it was a simple string.  i rewrote
the initial if form as a cond that will call ccl::ensure-simple-string
if the arg is indeed a string.

also updated ccl::new-package-name.