Wrap BEGIN-EDITING/END-EDITING around command processing.
The good news is that this suppresses display after every buffer
modification.

The bad news is that some extended/recursive commands have to
do hairy things to allow redisplay to happen.  (The hair mostly
has to do with closing the outer BEGIN-EDITING and re-opening
it to keep things balanced; sometimes this needs to happen
when reading characters from the event queue, since some types
of redisplay happen then.)

Summary: less hysterical redisplay code, but it may still be
suppressed too much; some things that don't really modify the
buffer may wrap a MODIFYING-BUFFER around a sequence of lower-
level operations, and this may cause the buffer to be marked
as modified even when nothing changes.