If we handle a SIGNAL_FOR_PROCESS_INTERRUPT trap, clear tcr->interrupt_pending.
Back out of the hack that tried to work around a Darwin bug (FP exception
flags were cleared in the MSR after a thread had been interrupted via
pthread_kill()), since the fix seems not to be GC-safe.

Add some debugging code to print when the exception lock is acquired/released.