Use #_ _exit for fatal termination.
Try to run lisp code after #_exit via lazarus(); this is for some
cases of SAVE-APPLICATION (e.g., in the Cocoa GUI).

The environment in which The Undead code runs is a little funky,
especially on PPC.  There might be GC implications (cstack will
contain call history as of the point when #_exit called atexit
handlers, vstack & tstack are emptied in lazarus()).  We should
barely run any code in this environment, but we will GC a time
or two ...