Changes to GUESS-ALIGNMENT.  Again, this only makes sense for cases where we have to guess; a bitfield that's 8 bits wide still has an alignment of 1.
Factor in the alignment of the containing int when bitfields are
involved.

When parsing a record type, set the "bits" slot of each field.
(Not too many things look at that slot, but we want the effect
of loading a record from the database to match the effect of
parsing a definition as exactly as possible, and to be able to
use EQUALP to compare the results.)