Revert out of some of the changes that led to Trac bug #2.
Try hard to preserve identity of named foreign-record-types; check for field
redefinition earlier.

Use PARSE-FOREIGN-TYPE in load form for FOREIGN-RECORD-TYPE, so that we
go through the same interning (of named types) and field compatibility
checking.  Move some of the FOREIGN-RECORD-TYPE definitions later in
the load order, since it's never really worked (doesn't enhance cross-compilation
to define them inside INSTALL-STANDARD-FOREIGN-TYPES, and since we now need
to be able to call PARSE-FOREIGN-TYPE, which can't work that early.
(Actually, it might work to call PARSE-FOREIGN-TYPE to define structure types
at load time, rather than dealing with the strange type constants introduced
in the expansion of DEF-FOREIGN-TYPE.)