In DECODE-ARGUMENTS (used to decode "regular" function arguments/return value), make sure that a foreign typedef name is escaped.
When decoding ObjC method arguments/result types, tell %DECODE-TYPE
not to expand typedefs.  (This helps to canonicalize things if the
typedef refers to an anonymous type, e.g., the :<NSD>ecimal case
in the ObjC bridge.)

In ENCODE-FFI-ARG-TYPE, handle return values the same as arguments.
Don't expand typedefs, just encode the name in the signature.

%DECODE-TYPE takes an extra SUPPRESS-TYPEDEF-EXPANSION argument.