Implement "Execute Selection" item on the Lisp menu, with shortcut command-E.
It's at least semi-standard for command-E to be used for
"Use Selection for Find", but in a nod to MCL muscle memory,
we'll usurp it.

cocoa.lisp: added #/ensureListener: action (and minor update to
#/showListener:).

cocoa-listener.lisp: call #/ensureListener: on the main thread from
ui-object-choose-listener-for-selection so that we always have a
listener to send data to.

cocoa-editor.lisp: add menu item validation for #/evalSelection:.

(As a bonus, changed OpenMCL Help... in the Help menu
to Clozure CL Help...)