Speed up conflict checking in snap-reader-methods, and fix a typo that made it misreport the number of conflicts as 0.
If ccl::**continue-from-readonly-hashtable-lock-error* is true, the
error on attempting to modify a read-only hash table will be
continuable, and continuing will make the hash table writeble. The
default for ccl::*continue-from-readonly-hashtable-lock-error* is nil,
so you have to set it to true in order to see the continue option, and
then you get BOLD CAPS AND EXCLAMATION POINTS IN YOUR FACE!

I actually ran into this just now, and boy was it annoying.