Got rid of ccl:cocoa-ide;Clozure CL.app and moved its contents to ccl:cocoa-ide;ide-contents;
Modified the IDE loading scheme to create bundles by copying from
cocoa-ide;ide-contents; as follows:

(require'cocoa) creates 'ccl:temp bundle.app', a fake bundle that is only
there to enable using cocoa immediately after loading it.  This is not
something that users are expected to do in normal circumstances, but is just
there for us to bootstrap.  (Users who use cocoa will ordinary use the IDE,
which will have cocoa preloaded).

(require'cocoa-application) creates 'ccl:Clozure CL.app', the real IDE bundle
that is ready to use and/or redistribute.  By default it includes the
interface headers for the target architecture, so it can be moved out of the
ccl directory if so desired.  I also made it load easygui unconditionally.