Implement symbol completion, and bind it to meta-/ (as in Emacs and Lispworks) and meta-` (as in MCL).
(defcommand "Expand Dynamic Abbreviation" (p)
  "Treats the symbol before point as an abbreviation and expands it.
It checks the following in order until a suitable expansion is found:
  - last accepted expansion for this abbreviation, if any
  - symbols in current buffer before point
  - symbols in current buffer after point
  - symbols in all other editor windows, front to back
  - symbols visible in the current package, fbound/bound symbols first
  - symbols in all other packages (in no particular order)

If called repeatedly from the same position, replaces the previous expansion
with the next possible one.

A symbol is a suitable expansion for an abbreviation if the abbreviation is
a proper prefix of the symbol, or the abbreviation consists of the initials
of the individual words within the symbol (e.g. mvb => multiple-value-bind).
"
  (declare (ignore p))