Optimize generic functions with lots of same-type EQL methods by using a hash table intead of an alist to look up the EQL arg.
Fix a bug in the code that prevented two threads from overwriting the
same generic function dispatch table entry at the same time. This bug
could cause redefining a method to forevermore slow down generic
function dispatch on arguments that hashed to the same place in its
generic function dispatch table. The effective method would never be
cached.

There's still some room for optimization of EQL methods. We could use
an array instead of an alist, which would cut the number of memory
references in half and make them more local. If all the methods on a
generic function are EQL methods on values of the same type, instead
of going through %%1st-arg-dcode to get to the EQL combined method, we
could go directly to the EQL combined method.

%gf-dispatch-table-store-conditional overwrites nil or
*gf-dispatch-bug*, not just nil.

grow-gf-dispatch-table uses *gf-dispatch-bug*, not another identical
in-line anonymous function.

compute-eql-combined-method-hash-table-threshold helps in determining
at which point it's faster to use a hash table than an alist for EQL
method dispatch. There are no callers of this function, but it's there
to aid determination of the value for
*eql-combined-method-hash-table-threshold*.

make-eql-combined-method now creates a hash table and uses
%%hash-table-combined-method-code if there are more than
*eql-combined-method-hash-table-threshold* methods.

%%hash-table-combined-method-dcode works the same as
%%assq-combined-method-dcode and %%assoc-combined-method-dcode, but
with a hash table instead of an alist as the dictionary that maps
the EQL arg to its effective method.