Add :monitor-futex-wait to *known-optional-features*.
RECURSIVE-LOCK-WHOSTATE: get the whostate field, not the value field ...

Deal with whostate a little differently in FUTEX-WAIT, LOCK-FUTEX, and 
callers.

Conditionally (#+monitor-futex-wait) count futex-wait syscalls and the
time spent in them in CCL::*TOTAL-FUTEX-WAIT-CALLS* and
CCL::*TOTAL-FUTEX-WAIT-TIMES*, respectively.  (This isn't thread-safe:
two threads trying to increment these values at the same time can cause
the results to be inconsistent.)