More source location optimizations (and one bug fix)
Moved the binding of *definition-source-note* to compile-named-function. This ensures that when emitting source notes 
the current value *definition-source-note* is the same one for the code we're compiling (it was not before). This 
change also almost halves the number of times we call x862-generate-pc-mapping since we no longer call it for 
note-function-info forms generated by defuns.

Merged the helper functions directly into x862-generate-pc-source-map.

Removed the unused generic functions source-note-text and (setf source-note-text).

There is still a bug in substream, or somewhere, where we're getting garbage (#\^Z chars) in our text streams, made 
%fast-compress ignore these chars.