Make compiler use CISC-y get/set/test bit instructions when referencing/setting bits in a bit vector.
In order to avoid minor bootstrapping problems, define a few new
bit referencing and bit testing vinsns with an "n" prefix, leaving
the old ones around for now.

The x86 BT/BTS/BTR instructions take two operands: a bit base and
bit offset.  When the bit base is a memory location, the bit offset
(in a register) is a signed 16/32/64 bit quantity (depending on
operand size).  There's no need to limit the bit offset to
(1- TARGET::NBITS-IN-WORD).

This is shorter and removes the need to use a register for a word index.
(The register saving is no big deal on x86-64, but it matters a lot on
32-bit x86.)