level-0/l0-io.lisp: Ensure that blocking I/O syscalls (fd-read, fd-write) ignore EINTR.  (If syscalls aren't restartable, then interrupted syscalls can return EINTR.  If we're trying to interrupt something via PROCESS-INTERRUPT, we want the syscall to exit the kernel after the handler runs, so that lisp code can notice the pending interrupt.)
lib/macros.lisp:
Try to make sure that the compille-time processing for DEFMETHOD
notes the method's lambda list as if &ALLOW-OTHER-KEYS was
specified whenever &KEY was.  (It may not be that wise to
base our notion of a function's arglist on the latest method
we've seen; if this doesn't work for all cases where this branch
matters, we might just note the DEFGENERIC arglist at compile-time
and ignore the arglists of individual DEFMETHODs.)

lib/pathnames.lisp:
Add an obscure argument to DIRECTORY, default it to T.  DIRECTORY
will ignore files whose names start with ".#" (which are often
broken links that Emacs uses as lockfiles).  This means that people
who have real files whose names start with ".#" won't be able to
see them unless they override this option.  Or they could give
them less brain-dead names ...

lisp-kernel/x86-exceptions.c:
Don't set the SA_RESTART open when installing signal handlers.
(This change was made in the trunk, but hadn't been backported
to the working-0711 branch.)  The effect of the change is that
system calls (accept, read, write) that previously were 
automatically restarted whenever they're interrupted now return
an "I was interrupted" error (#$EINTR).  Returning from the OS
kernel (rather than restarting the syscall) means that pending
interrupts (sent via PROCESS-INTERRUPT) get handled on exit
from the syscall.  (If we return from the lisp handler, we have
the option of restarting the syscall.)

None of this is hard to bootstrap, but the changes that make
syscalls interruptible require running under a recompiled
kernel, and handling the case where a "write" syscall is interrupted
requires changes to the lisp code.  Do a full rebuild a time or two
and things should be fine.