This commit includes support for docbook 4.5, stylesheet changes, and updated documentation.
In order to support docbook 4.5 in nXML mode, I have added a new directory called docbook-rng-4.5 and changed schemas.xml to point to it.  This should just work when editing the documentation in EMACS.

The two most obvious changes to the stylesheets are that the table of contents for each chapter now occurs at the beginning of the chapter, and the format for refentries is cleaner and more concise.

I think that we should consistently use refentry elements for all of the definitions of functions, macros, variables, etc.  This retains the structured data for the definitions that can be reformatted to have different appearences by the stylesheets.  We should also consistently use other docbook elements such as function and varname.  I'm not really happy with their appearance right now, but that can be easily tweaked in the stylesheets as long as they are consistently used throughout the documentation.