Add SPgets64.
The thought at the moment is to use mm0 (rather than the %edx:%eax
pair) as the place where gets64/getu64/makes64/makeu64 put or get an
unboxed 64-bit quantity.

If this idea holds up, SPsyscall will to have to know when a syscall
returns a doubleword result, and copy the %edx:%eax pair into %mm0.