.SPbind_self, .SPbind_self_boundp_check, .SPmvstackv, .SPtfuncallgen, .SPunbind_n, .SPunbind_to: implement
Nasty kludge in .SPffcall to dump the x87 state to memory to see if there
is a value on the x87 stack, popping the x87 stack and putting the result
in xmm0 if so.  The compiler should take care of popping the x87 stack if
needed, since it will know whether an ffcall returns an fp result.  This
won't be hard to do, but a few x87 instructions need to be added to the
assembler first.