%%0-ARG-DCODE use *GF-PROTOT* trampoline, so register it.
REMVOE-ENCAPSULATION: if dcode isn't registered, default to
#'FUNCALLABLE-TRAMPOLINE.  

There's actually one other case, which involves a funcallable-instance
whose funcallable-instance-function is unset; the trampoline signals
an error to that effect.  I'm not sure if we can trace such a function;
if so, we might want to handle it in REMOVE-ENCAPSULATION.

I think that this may close ticket:287.