Add file makedmg to scripts.
This is a simple utility to create a disk image from the current directory.
The resulting dmg file is placed in the parent directory.

The volume name of the disk image is the final component
of the current directory name.
The file name of the image is the same with ".dmg" appended.