Do STRING-EQUAL differently.
Do INTERN, FIND-SYMBOL, FIND-PACKAGE, and CCL::PKG-ARG with a constant,
stringable package-name arg via a "package-refrence" (a little structure
that maps package names to packages.)  There's runtime support intended
to keep "package-referencess" up-to-date as packages are named/renamed,
so the transformed versions of these functions should do package-name
lookups in unit time.

Add compiler-macro. for (new) %CLASS-ORDINAL.

When SUBTYPEP's second argument is a constant type-specifier, use a 
TYPE-REFERENCE structure to map the type-specifier to a CTYPE.  Probably
needs more runtime support to handle redefinition of types.