Port the "extensible/non-dt dcode functions" stuff from the trunk.
In %CLASS-CELL-TYPEP, use the new cpl-bitmap.

Assign consecutive small integers ("ordinals") to clases; still no
support for foreign-class ordinals.  For standard classes, use 
instance.hash to store the ordinal.

Update a bitmap whenever the CPL is cached in the class-wrapper.

Make built-in istruct classes for package-ref, type-ref.

INSTANCE-CLASS-WRAPPER is now a non-generic function.