Use bitvectors in some of the optimized-reader-dcode functions.
Define *UNIQUE-READER-DCODE-FUNCTIONS*, defaults to T.  Some of the
optimized reader dcode functions might perform better if they're
compiled in a way that avoids using the dispatch table to pass
parameters.

INSTANCE-CLASS-WRAPPER is now a non-generic function.

Define REGISTER-NON-DT-DCODE-FUNCTION, use it to define specialized
dcode functions.  Define exactly one such function, which causes the
method-function to be called directly if (a) there's only one method
defined on a GF and (b) it's universally applicable.  (There are
a handful of such methods in CCL, over 100 in a customer's application,
but don't know how often they're called.)