Use (typep x 'hash-table) instead of (hash-table-p x).  Mostly because we can, but may be a little faster.
There's an (unresolved) issue related to the definition of type-predicates
on ISTRUCT classes that have no subclasses: in the TYPEP compiler-macro,
it's generally better to recognize the istruct class and do an inline
ISTRUCT-TYPEP than it is to call a function that does that; at compile-time,
we can certainly afford a FIND-CLASS.  At runtime (in #'TYPEP), calling
a predicate if it exists is likely to be much, much cheaper than going
through translation/%typep would be.  I started to remove type-predicates
on simple istruct classes before realizing this; it may be better to
keep them around and re-work OPTIMIZE-TYPEP to try the simple istruct
class case first.

In any case, (typep x 'hash-table-p) and (hash-table-p x) should ideally
be equivalent in all respects.