CHECK-SEQUENCE-BOUNDS has to signal TYPE-ERRORs in some cases.
New support for case folding: the maximum Unicode char-code for which
either CHAR-UPCASE or CHAR-DOWNCASE returns something other than its
arg is <70K.  Those codes wouldn't all fit in a vector with 16-bit
elements, but the maximum difference between any char-code and its
equivalent is ~10K (the vast majority of differences are actually
of type (SIGNED-BYTE 8), but there are some outliers.)  So, two
sparse ~70K-element vectors of (signed-byte 16) can map from any
character that can be case-folded to the difference between that
code and its equivalent, which means that we can do case folding
as comparison, aref, and add.

Update char-upcase/downcase and friends to exploit this; they
generally get faster as a result.