Try to exploit the fact that more "standard" slots are now present in CLASS and BUILT-IN-CLASS, so we can use simpler accessors in some cases.
Except for the OBJC-BRIDGE, CLASSP is a standard-instance whose
instance.hash slot contains an integer less than MAX-CLASS-ORDINAL.
(1M) (This currently assumes that non-class instances - whose
instance.hash slots contain a fixnum derived from the instance's
address - are never allocated in the low 1MB of the address space,
and that there will never be more than 1M classes; both seem to
be safe assumptions at the moment.

SET-FIND-CLASS tries to keep the class's notion of whether or not
it's "properly named" in synch.  (See also (SETF CLASS-NAME)).

ISTRUCT-based classes are made by a new primitive, which arranges
to ensure that the unique CONS used to map istruct instances to 
their class info is registered (and references the class wrapper.)

The CLASS-OF method for ISTRUCTs explots this.

The functions that we get to via SLOT-ID-VALUE - and which try
to decide whether or not to call [SETF] SLOT-VALUE-USING-CLASS -
recognize structure instances.

Provide runtime support for doing SLOT-BOUNDP via a SLOT-ID.

Try to speed up slot lookup for runtime calls to SLOT-VALUE and friend;
FIND-SLOTD is just a loop.

In ALLOCATE-INSTANCE on a sructure, put list of class-cells (not
class-names) in 0th element of structure.