Bump the fasl version.
Bring the list of istruct types that have MAKE-LOAD-FORM methods
defined on them up to date.

When we scan lists, try to recognize lists that're used as
"istruct cells".  (Having a separate typecode for istruct cells
would obviate the need for this, but it's probably not a huge
performance hit.)  When we dump lists, dump istruct cells with
a new fasl operator.