Do raise_thread_interrupt() for Win64.
If the thread is running foreign code when we interrupt it, we set a
bit in its tcr.  (This is the same as what happens on other
platforms.)  However, on Unix platforms (where this works via signals),
the delivery of a sigmal to the target thread causes any blocking
syscalls to return with EINTR (and thus usualy quickly qreturn to lisp glue
code that checks the interrupt-pending flag.)

If a thread's in an "alertable wait state", using QueueUserAPC might get
it out of that state.  I'm not sure how to tell whether or not the thread's
in such as state.