Avoid a few cases of partial register writes (assignments to the low 16 bits of a register), which can have a surprisingly severe negative effect on performance.  (Catching more cases involves some bootstrapping of kernel changes from the trunk, to guarantee that "set_nargs()" always zero-extends the argument count through all bits in the register.)
Make CODE-CHAR return NIL for codes #xFFFE and #xFFFF, which are defined
(along with things in the surrogate-pair range) to be invalid.  This
slows down CODE-CHAR a little (obviously), but avoids problems and
keeps higher-level things from having to worry about that.

(Note that (CODE-CHAR (THE CCL::VALID-CHAR-CODE x)) skips this checking
when it's safe to do so.)