Define the SERIOUS-CONDITION EXTERNAL-PROCESS-CREATION-FAILURE.
Add *SILENTLY-IGNORE-CATASTROPHIC-FAILURE-IN-RUN-PROGRAM, initialize
it to T on this branch.

Add :SILENTLY-IGNORE-CATASTROPHIC-FAILURE arg to RUN-PROGRAM.  When
true, an EXTERNAL-PROCESS with status :ERROR and some exit code
(and a null PID) is returned, as if that was a reasonable thing
to do.  If catastrophic failure (#_fork error) isn't silently
ignored, a serious condition of type EXTERNAL-PROCESS-CREATION-FAILURE
is signaled.

Make things that expect EXTERNAL-PROCESS-PID to be non-NIL handle
the possibilit that catastrophic falures have been ignored.