Lots of bugs (many long-standing) exposed by test suite; we're now running it with a policy that allows compiler-macros to be used.
INFER-ARRAY-TYPE: in general, the best we can do for a non-constant
element-type is *, but that's better than treating non-constants
as type-specifiers.

CONSTANT-KEYWORDS-P does (as it apparently once did) need to check
for duplicates; we can't do destructuring over a list of keyword/
value pairs unless we're sure that there are no duplicates, since
the values associated with the duplicate keys might have side-effects.
(If we're sure that any such values are constants, we can allow them
as long as COMP-NUKE-KEYS does the right thing (e.g., uses the value
associated with the leftmost instance of the keyword.)

COMP-NUKE-KEYS: don't set positional arg in "call-list" if it has
already been set (e.g., if duplicate keywords, leftmost wins.)

If TYPEP transforms to a call to ARRAY-%%TYPEP, return just the first
value.

Character comparisons involving 3 args have to evaluate the third
form (for side-effects), even if the function result can be determined
by looking at the first 2 characters.  (The way that this is done,
(char< #\b #\a 17) might signal an error when inlined and the function
may not, but that case is bogus, anyway.)

These bugs caused about 180 test failures; a few more tests (involving
bignum/float comparisons and RANDOM) still fail, but I haven't yet
been able to get them to fail in an environment where I can reproduce
the failure.