Lots of changes coming on this branch; will check in binaries when all done.
Executive summary: merge many changes from trunk (making trunk and
this branch closer), change the address of NIL and other static
constants to be above the limit (vm.mmap.min_addr) set in recent
Linux distributions.  Change fasl version and image-abi-version,
so code compiled in previous versions won't work in this version
(and vice versa), mostly due to the different notions of the
values of NIL and other constants.

For these files: copy versions from trunk as of this date.