Merge changes from branches/win64.
As well as the expected low-level exception/suspend/interrupt stuff,
these changes also include changes to [f]printf format strings.  Note
that on win64, a 'long' is 32-bits wide, which complicates matters:

- an address (viewed as an integer) or a natural-sized integer isn't
(portably) a long, and so can't be printed with %l.

- an address (viewed as an integer) or a natural-sized integer isn't
(portably) a long long, and so can't be printed with %ll.

- an address (viewed as an integer) or a natural-sized integer can be
portably printed with '%p', but implementations differ as to whether
or not '%p' prepends a gratuitous '0x' to the hex address.  (Linux
does, other current platforms seem not to.)

The approach that seems to work is to cast arguments to natural, then
to u64_t, then use %ll.  That approach probably isn't taken consistently
(yet), so some debugging information printed by the kernel may be
incorrect.