Assorted code coverage extensions.
- added an :HTML argument to CCL:REPORT-COVERAGE, can be specified as
nil to skip writing html files.

- New function CCL:COVERAGE-STATISTICS - returns a sequence of
CCL:COVERAGE-STATISTICS objects containing the same information that
is written to the statistics file by report-coverage.  See doc (or
source) for the list of accessors.

- Make coverage snapshots be first class objects.  New functions
related to snapshots:

 CCL:SAVE-COVERAGE - creates a snapshot
 CCL:RESTORE-COVERAGE - restores the coverage from a snapshot
 CCL:COMBINE-COVERAGE - combines a set of snapshots into one
 CCL:WRITE-COVERAGE-TO-FILE - saves a snapshot in a file
 CCL:READ-COVERAGE-FROM-FILE - loads a snapshot from a file

- Changed the internal representation so that sparse snapshots
(i.e. when most code is not covered) take less space, in anticipation
of having thousands of them around.

- Made the file written by write-coverage-to-file (and hence
save-coverage-to-file as well) be a lisp source file, which can be
compiled for possibly faster loading.  Either the original file or
the compiled file can be handed to read-coverage-from-file or
restore-coverage-from-file.

The documentation in http://trac.clozure.com/openmcl/wiki/CodeCoverage
has been updated with all these changes.