In PRINT-UNREADABLE-OBJECT.2, mark the form that tests for output of the form #<SOME-TYPE > (note space) as bogus.
I suppose I can see how a pedantic reading of the spec could lead to  
this conclusion, but since the spec also implies that the precise  
format of stuff between "#<" and ">" is implementation-dependent, it's  
hard for me to agree.  It's not like programs can portably rely on the  
output of PRINT-UNREADABLE-OBJECT anyway.  In the case above, it just  
looks stupid.  (I mean, if it's valid to omit the body forms, then it  
stands to reason that it's valid to omit the extra spaces that would  
normally delimit the output from the body forms.  Sheesh.)