Define c_stack_16_byte_aligned.  Not sure why I did it this way rather than in m4.
It's not clear whether x8632 Linux tries too hard to maintain
16-byte stack alignment; it may be that main() forces the stack
to be aligned, but thread startup code on other threads doesn't
care.  The ABI document from the mid 90s doesn't require anything
more than word alignment.

In any case, it's often not aligned right on Linux in practice,
so use this constant to conditionalize.