In windows_arbstack_exception_handler: insist that we're running on some stack other than the C stack when the exception occurs (in general, we aren't going to be able to do much about exceptions that happen on the C stack, anyway.)  Always build a (copy of) the exception context/ info on the foreign stack; don't switch to the foreign stack from the handler, but setup the exception context we were called with to enter the higher-level exception handler when it's resumed, and tell the OS to resume execution.  (While we're at it, arrange that the higher-level handler gets the TCR as an argument, to avoid having to look it up again.)
This allows the Windows runtime to clean up after itself after
returning from the low-level handler (releasing locks, etc.), and has
the effect of calling our higher-level handler on the foreign stack
with a copy of the original exception context/info.

(If there was ever any reason to, we could do the stack switch the
same way on other platforms.  On non-Mach systems, I think that the
handler is called directly from the kernel and on Darwin we generally
fake things using Mach exception-handling futilities, so there doesn't
seem to be any need for this elsewhere.)