Extend the mechanism used for keeping track of definitions (previously used only for inlining and undefined function warnings) to make it work across compilation units rather than just per file and make it also keep track of method definitions.  Make it use a hash table since now the set of definitions can get quite large. Use it to detect and issue warnings about duplicate function or method definitions inside a compilation unit.
One side-effect of above changes is that compile-time inlining information is kept across the whole compilation unit rather than being per-file only.

Fixes in defstruct, to make sure predicate def gets noted, and to make accessor defs only get noted once.

Add (and export) a new CCL:WITHOUT-DUPLICATE-DEFINITION-WARNINGS macro that can be wrapped around intentional duplicate definitions to avoid the warning.

New NLAMBDA macro, which is to LAMBDA as NFUNCTION is to FUNCTION.  Not exported.

Wrap with-compilation-unit around compile-ccl and xcompile-ccl.

Remove a half-dozen or so unintentional duplicate definitions in ccl, and work around the intentional ones by either using fset instead of defun at the first definition or wrapping the second one in without-duplicate-definition-warnings.

In file compiler, do not use the file's lexical environment for compile-time evaluation!

Make report-compiler-warning show the file position, when known, for warnings in anonymous lambdas.