In the Windows version of terminate_lisp(), use #__exit (not ExitProcess, which seems to want to run things registers via #_atexit.  Windows' little joke.)
Don't claim to run on Panther on Darwin.  (We probably haven't in quite
a while, but try to catch that earlier.)

Use SPJUMP_TARGET_ADDRESS in all of the various versions of remap_spjump;
in the Windows version, consider the possibility that the jump table's
already where it needs to be.