From working-0711 branch: add new fn, not used by anything yet:
READ-RECORDING-SOURCE (stream &key eofval file-name start-offset map)
Read a top-level form, perhaps recording source locations.
If MAP is NIL, just reads a form as if by READ.
If MAP is non-NIL, returns a second value of a source-note object describing the form.
In addition, if MAP is a hash table, it gets filled with source-note's for all
non-atomic nested subforms.