Source location support in the compiler:
COMPILE-NAMED-FUNCTION takes a new SOURCE-NOTES arg, which should be nil or a hash table mapping source forms to source notes.  In the latter case, the compiler will do its best to track the source notes from the source all the way through code generation, and create a pc/source map, storing it as the 'pc-source-map property on the %lfun-info plist of the function and any inner functions.  In addition, the compiler will store the source note of the lambda form on the 'function-source-note property of the function and any inner functions.

COMPILE-NAMED-FUNCTION also takes a new FUNCTION-NOTE arg which can be used to override the lambda source note indicated by SOURCE-NOTES.

Nothing actually passes in either of these arguments yet.


Also checking in some cases of acode-unwrapped-form -> acode-unwrapped-form-value, which have nothing to do with source locations but just help minimize diffs for easier merging.