X862-VAR-NOT-SET-BY-FORM-P and X862-SETQED-VAR-NOT-SET-BY-FORM-P: the mysterious SIDE-EFFECT-FREE operator bit basically means "doesn't do SETQ or cause SETQ to happen in mysterious ways", e.g., in cases involving closures.  Only insist on that bit if the var is closed over.
In X862-VREF1, pay attention to vreg more for FPRs, too.  Handle some
cases (now 64-bit only) where we have a constant fixnum index that
doesn't fit in a 32-bit displacement and we're being unsafe (the
caller doesn't set UNSCALED-IDX in that case, see ticket:375.)

In X862-VSET: at least on x8664, when no need to memoize, can use
"untargeted" registers.

Pushes on/pops from csp: use x862-open-undo rather than trying
to track the exact effects on cstack.  (Not clear if we can clear
N words off of cstack, and it's not clear that N was right here,
either.)

X862-ACC-REG-FOR: chose arg_z if reg is a node reg, otherwise just
use it.  (It's likely that whatever register is involved here is
just going to be pushed, so we really want to say "whatever's
easiest.")

Handle natural comparison a little differently, especially on ia32.
We can't easily use a general targeting mechanism here to steal
an extra imm reg here, since we may want to branch out of the scope
of and beyond the extent of the mark/unmark.